requirements.txt
streamlit==1.24.0
langchain==0.0.200
openai==0.27.8
python-dotenv==1.0.0
chromadb==0.3.26
pandas==2.0.3
sqlite3==3.42.0
requests==2.31.0